BIA/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.3
pillow==10.2.0
proglog==0.1.10
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing==3.7.4.3
tzdata==2023.4
urllib3==2.1.0
